% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wege_function.R
\name{get_wege}
\alias{get_wege}
\title{get_wege}
\usage{
get_wege(
  target_area,
  input,
  x,
  y,
  species = "binomial",
  category = "category",
  res = 1
)
}
\arguments{
\item{target_area}{Either a sp or sf object to which to calculate the WEGE index.}

\item{input}{Species ranges, either from a shapefile or from a georeferenced species list with a column for species, two columns for coordinates and one for 
the IUCN category.}

\item{x}{name of the longitude column.}

\item{y}{name of the latitude column.}

\item{species}{name of the species column.}

\item{category}{name of IUCN the category column. Terminology must be as follows: DD for Data Deficient, LC for Least Concern, NT for Near Threatened, 
EN, for Endangered, CR for Critically Endangered, EW for Extinct in the wild 
and EX for Extinct.}

\item{res}{grid-cell size to use to calculate the range of the species in 

case a georeferenced species list was provided.}
}
\description{
A function to get the WEGE index value for a provided polygon.
}
\examples{
examples/get_wege_function.R

}
