% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_ras_function.R
\name{spat_ras}
\alias{spat_ras}
\title{spat_ras}
\usage{
spat_ras(
  target_area,
  input,
  x,
  y,
  species = "binomial",
  category = "category",
  ed = FALSE,
  res = 1
)
}
\arguments{
\item{target_area}{Either a sp or sf object to which to calculate the WEGE index.}

\item{input}{Species ranges, either from a shapefile or from a georeferenced species list with a column for species, two columns for coordinates and one for the IUCN category.}

\item{x}{name of the longitude column.}

\item{y}{name of the latitude column.}

\item{species}{name of the species column.}

\item{category}{name of IUCN the category column. Terminology must be as follows: DD for Data Deficient, LC for Least Concern, NT for Near Threatened, EN, for Endangered, CR for Critically Endangered, EW for Extinct in the wild and EX for Extinct.}

\item{ed}{name of the evolutionary distinctiveness column.}

\item{res}{grid-cell size to use to calculate the range of the species in case a georeferenced species list was provided.}
}
\description{
A function to get the WEGE index value for a provided polygon.
}
\examples{
get_wege(target_area,input,species = 'binomial',category = 'category')
get_wege(target_area,input,x = 'decimallongitude',y = 'decimallatitude',species = 'BINOMIAL',category = 'category')



}
