% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WEGE.R
\docType{package}
\name{WEGE}
\alias{WEGE}
\title{WEGE: Describe the package here in more detail! CRAN will complain if you
don't.}
\description{
WEGE: Describe the package here in more detail! CRAN will complain if you
don't.
}
