% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WEGE.R
\docType{package}
\name{WEGE}
\alias{WEGE}
\title{WEGE: A new metric for ranking locations for biodiversity conservation}
\description{
WEGE is an R package that allows the user to calculate the WEGE index for a particular area. Additionally it also calculates rasters of KBA criteria (A1a, A1b, A1e, and B1) Weighted endemism, the EDGE score, Evolutionary Distintiveness and Extintion risk.
}
